polarimetry: move all dualbeam things to __init__
